Fixed for kotlin 1.9

+++ ktor-client/ktor-client-winhttp/windows/src/io/ktor/client/engine/winhttp/internal/WinHttpRequest.kt
@@ -30,7 +30,7 @@ internal class WinHttpRequest(
     private val connectReference: StableRef<WinHttpConnect>
 
     init {
-        val hConnect = WinHttpConnect(hSession, data.url.host, data.url.port.convert(), 0)
+        val hConnect = WinHttpConnect(hSession, data.url.host, data.url.port.convert(), 0U)
             ?: throw getWinHttpException("Unable to create connection")
         connect = WinHttpConnect(hConnect)
         connectReference = StableRef.create(connect)
@@ -58,7 +58,7 @@ internal class WinHttpRequest(
     }
 
     fun upgradeToWebSocket() {
-        if (WinHttpSetOption(hRequest, WINHTTP_OPTION_UPGRADE_TO_WEB_SOCKET, null, 0) == 0) {
+        if (WinHttpSetOption(hRequest, WINHTTP_OPTION_UPGRADE_TO_WEB_SOCKET.toUInt(), null, 0U) == 0) {
             throw getWinHttpException("Unable to request WebSocket upgrade")
         }
     }
@@ -86,8 +86,8 @@ internal class WinHttpRequest(
                     headersString,
                     headersString.length.convert(),
                     WINHTTP_NO_REQUEST_DATA,
-                    0,
-                    WINHTTP_IGNORE_REQUEST_TOTAL_LENGTH,
+                    0U,
+                    WINHTTP_IGNORE_REQUEST_TOTAL_LENGTH.toUInt(),
                     statePtr.convert()
                 ) == 0
             ) {
@@ -221,7 +221,7 @@ internal class WinHttpRequest(
 
         if (WinHttpSetOption(
                 hRequest,
-                WINHTTP_OPTION_DISABLE_FEATURE,
+                WINHTTP_OPTION_DISABLE_FEATURE.toUInt(),
                 options.ptr,
                 sizeOf<DWORDVar>().convert()
             ) == 0
@@ -241,7 +241,7 @@ internal class WinHttpRequest(
             null
         }
 
-        val oldStatusCallback = WinHttpSetStatusCallback(hRequest, callback, notifications, 0)
+        val oldStatusCallback = WinHttpSetStatusCallback(hRequest, callback, notifications, 0U)
         if (oldStatusCallback?.rawValue?.toLong() == WINHTTP_INVALID_STATUS_CALLBACK) {
             val errorCode = GetLastError()
             if (errorCode != ERROR_INVALID_HANDLE) {
@@ -275,7 +275,7 @@ internal class WinHttpRequest(
                     SECURITY_FLAG_IGNORE_CERT_DATE_INVALID
                 ).convert()
         }
-        if (WinHttpSetOption(hRequest, WINHTTP_OPTION_SECURITY_FLAGS, flags.ptr, UINT_SIZE) == 0) {
+        if (WinHttpSetOption(hRequest, WINHTTP_OPTION_SECURITY_FLAGS.toUInt(), flags.ptr, UINT_SIZE) == 0) {
             throw getWinHttpException("Unable to disable TLS verification")
         }
     }
